+++ src/data/initialDatas.ts
@@ -1,0 +1,16 @@
+import { MenuListDataType } from '@/types/Menu';
+
+export const menuListData: MenuListDataType[] = [
+  {
+    id: 1,
+    href: '/events',
+    title: '기획전',
+    desc: '진행중인 기획전을 만나보세요.',
+  },
+  {
+    id: 2,
+    href: '/best',
+    title: '베스트',
+    desc: '스타벅스 베스트 MD 상품만 모아보세요.',
+  },
+];
