+++ src/types/Menu/index.ts
@@ -1,0 +1,6 @@
+export type MenuListDataType = {
+  id: number;
+  href: string;
+  title: string;
+  desc: string;
+};
